build: rename build/ to scripts/
