+++ package.json
@@ -13,16 +13,16 @@
     "types/*.d.ts"
   ],
   "scripts": {
-    "dev": "rollup -w -c build/config.js --environment TARGET:web-full-dev",
-    "dev:cjs": "rollup -w -c build/config.js --environment TARGET:web-runtime-cjs",
-    "dev:esm": "rollup -w -c build/config.js --environment TARGET:web-runtime-esm",
+    "dev": "rollup -w -c scripts/config.js --environment TARGET:web-full-dev",
+    "dev:cjs": "rollup -w -c scripts/config.js --environment TARGET:web-runtime-cjs",
+    "dev:esm": "rollup -w -c scripts/config.js --environment TARGET:web-runtime-esm",
     "dev:test": "karma start test/unit/karma.dev.config.js",
-    "dev:ssr": "rollup -w -c build/config.js --environment TARGET:web-server-renderer",
-    "dev:compiler": "rollup -w -c build/config.js --environment TARGET:web-compiler ",
-    "dev:weex": "rollup -w -c build/config.js --environment TARGET:weex-framework",
-    "dev:weex:factory": "rollup -w -c build/config.js --environment TARGET:weex-factory",
-    "dev:weex:compiler": "rollup -w -c build/config.js --environment TARGET:weex-compiler ",
-    "build": "node build/build.js",
+    "dev:ssr": "rollup -w -c scripts/config.js --environment TARGET:web-server-renderer",
+    "dev:compiler": "rollup -w -c scripts/config.js --environment TARGET:web-compiler ",
+    "dev:weex": "rollup -w -c scripts/config.js --environment TARGET:weex-framework",
+    "dev:weex:factory": "rollup -w -c scripts/config.js --environment TARGET:weex-factory",
+    "dev:weex:compiler": "rollup -w -c scripts/config.js --environment TARGET:weex-compiler ",
+    "build": "node scripts/build.js",
     "build:ssr": "npm run build -- web-runtime-cjs,web-server-renderer",
     "build:weex": "npm run build -- weex",
     "test": "npm run lint && flow check && npm run test:types && npm run test:cover && npm run test:e2e -- --env phantomjs && npm run test:ssr && npm run test:weex",
@@ -37,10 +37,10 @@
     "flow": "flow check",
     "sauce": "karma start test/unit/karma.sauce.config.js",
     "bench:ssr": "npm run build:ssr && node benchmarks/ssr/renderToString.js && node benchmarks/ssr/renderToStream.js",
-    "release": "bash build/release.sh",
-    "release:weex": "bash build/release-weex.sh",
-    "release:note": "node build/gen-release-note.js",
-    "setup": "node build/setup.js",
+    "release": "bash scripts/release.sh",
+    "release:weex": "bash scripts/release-weex.sh",
+    "release:note": "node scripts/gen-release-note.js",
+    "setup": "node scripts/setup.js",
     "commit": "git-cz"
   },
   "repository": {
