+++ scripts/gen-release-note.js
