+++ scripts/get-weex-version.js
