+++ scripts/release-weex.sh
