+++ scripts/setup.js
